refactor(evaluation): keep metric imports lightweight
