+++ tests/metrics/test_manifold_brier_per_agent.py
@@ -10,7 +10,7 @@ from datetime import UTC, datetime, timedelta
 
 import pytest
 
-from aragora.metrics.manifold_brier import (
+from aragora.evaluation.manifold_brier import (
     ManifoldBrierScorer,
     ManifoldPrediction,
 )
